requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
python-multipart
python-jose
passlib[bcrypt]
cryptography
azure-cognitiveservices-speech
openai
python-dotenv
pandas
openpyxl
pymupdf
fpdf
requests
PyJWT
